Add Nonlinearity-Distortion Plugin
Right now without nice oversampling. But for extra CPU usage, we
actually oversample and discard the restults :D  All the niceness of
oversampling when it comes to CPU and RAM usage but without any
benefits! ;) (This is the way it sounds best to my ears, something in
the up/down-sampling is definately wrong.)