Add gmSynth (NOT WORKING)
add libgig_based_dls_parser (only for windows, with vs2015 solution), add gm_synth and related classes without enough comments at this point. The synth compiles and wokrs but the sounds are not right (at all). I guess that we use the windows handle wrong or load the file at the wrong byte positions or cast wrong to uint16_t.